requirements.txt
openai>=0.27.0
python-dotenv
PyGithub
